feat : Navigation, Card, Applicant 로직 수정

+++ src/server/src/main/java/com/econovation/recruit/domain/board/NavigationRepository.java
@@ -6,7 +6,7 @@ import org.springframework.stereotype.Repository;
 import java.util.Optional;
 
 @Repository
-public interface NavigationRepository extends JpaRepository<Navigation, Long> {
+public interface NavigationRepository extends JpaRepository<Navigation, Integer> {
     Optional<Navigation> findById(Integer id);
     Optional<Navigation> findNavigationByNavLoc(Integer navLoc);
 }
